stop sw_emu and hw_emu jobs from entering AFI stage

+++ polyphemus/worker_f1.py
@@ -55,6 +55,10 @@ def stage_afi(db, config):
     (Xilinx FPGA binary file).
     """
     with work(db, state.AFI_START, state.AFI, state.HLS_FINISH) as task:
+        # sw_emu and hw_emu do not require AFI
+        if task['mode'] != 'hw': 
+            task.log('skipping AFI stage for {}'.format(task['mode']))
+            return
         # Clean up any generated files from previous runs.
         task.run(
             ['rm -rf to_aws *afi_id.txt \
